fix filename typo
